[lexical-react][lexical-playground] sync draggable block plugin to www (#6397)
