+++ packages/lexical-react/flow/LexicalDraggableBlockPlugin.js.flow
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+declare export function DraggableBlockPlugin(): React$MixedElement;
