+++ packages/lexical-react/src/shared/point.ts
